@@ -15,7 +15,7 @@ import fr.olympa.api.customevents.OlympaPlayerLoadEvent;
 import fr.olympa.api.groups.OlympaGroup;
 import fr.olympa.api.objects.OlympaPlayer;
 import fr.olympa.api.provider.AccountProvider;
-import fr.olympa.api.utils.SpigotUtils;
+import fr.olympa.api.utils.ColorUtils;
 
 public class SmallDataManagmentListener implements Listener {
 
@@ -65,7 +65,7 @@ public class SmallDataManagmentListener implements Listener {
 		Player player = event.getPlayer();
 		OlympaPlayer olympaPlayer = AccountProvider.get(player.getUniqueId());
 		if (olympaPlayer != null) {
-			event.setQuitMessage(SpigotUtils.color("&7[&c-&7] %prefix%name"
+			event.setQuitMessage(ColorUtils.color("&7[&c-&7] %prefix%name"
 					.replaceAll("%group", olympaPlayer.getGroup().getName())
 					.replaceAll("%prefix", olympaPlayer.getGroup().getPrefix())
 					.replaceAll("%name", player.getName())));
