@@ -23,11 +23,12 @@ public class OlympaPlayerObject implements OlympaPlayer {
 	private long id;
 	private OlympaGroup group;
 
-	public OlympaPlayerObject(UUID uuid, String name, String ip, long id) {
+	private OlympaPlayerInformations cachedInfos = null;
+
+	public OlympaPlayerObject(UUID uuid, String name, String ip) {
 		this.uuid = uuid;
 		this.name = name;
 		this.ip = ip;
-		this.id = id;
 	}
 
 	@Override
@@ -139,18 +140,16 @@ public class OlympaPlayerObject implements OlympaPlayer {
 
 	@Override
 	public long getId() {
-		// TODO Auto-generated method stub
 		return id;
 	}
 
 	@Override
 	public OlympaPlayerInformations getInformation() {
-		return null;
+		return cachedInfos == null ? cachedInfos = AccountProvider.getPlayerInformations(id) : cachedInfos;
 	}
 
 	@Override
 	public String getIp() {
-		// TODO Auto-generated method stub
 		return ip;
 	}
 
@@ -305,8 +304,7 @@ public class OlympaPlayerObject implements OlympaPlayer {
 
 	@Override
 	public void setId(long id) {
-		// TODO Auto-generated method stub
-
+		this.id = id;
 	}
 
 	@Override
