@@ -50,7 +50,7 @@ public enum Prefix {
 	}
 
 	public void sendMessage(CommandSender sender, String msg) {
-		sender.sendMessage(SpigotUtils.color(prefix + msg));
+		sender.sendMessage(ColorUtils.color(prefix + msg));
 	}
 
 	private void setPrefix(final String prefix) {
@@ -59,7 +59,7 @@ public enum Prefix {
 
 	@Override
 	public String toString() {
-		return SpigotUtils.color(prefix);
+		return ColorUtils.color(prefix);
 	}
 
 	public String toStringWithoutFormat() {
